Divisão dos testes para diminuir a complexidade
